Patch issue with share plans on development
Devise circumvents running validations on the user models when invite! is called.
I updated this call to adhere to validations, and I've updated the form to request the name,org of the user when a plan is shared with them.
This should fix the issue for both newly created share-plan users, and those with previous, bad data in the system.

You may want to consider re-running your validations(or fixup) scripts for users as you can have some invalid ones that you'll need to set a default name/org