Fix assets precompile task adding nulldb adapter
The rake task was breaking because it was trying to initialize the
database. The database is not available when building the RPM for
deployment.

Using the nulldb database adapter addresses this issue when precompiling
with the command:

```
DATABASE_URL=nulldb://user:test@127.0.0.1/roadmap RAILS_ENV=uat bundle exec rake assets:precompile
```