Add initial work for translation.io for Theme
This is an initial step towards translating the Theme titles.

The rake task roadmap_translation:export_theme_values fetches the titles
from Theme in the database and writes them to an external view which is
only used to feed the values to translation.io . After this is done we
need to sync the values as always with rake task translation:sync .

This commit also changes the way Theme titles are presented and sorted
using the I18n translations.