DCC Issue 710: The V0 API calls for assign_users and unassign_users to departments was broken.
DCC Issue https://github.com/DigitalCurationCentre/DMPonline-Service/issues/710

The fix involved making the following changes to the assign_users() and
unassign_users() methods in
app/controllers/api/v0/departments_controller.rb:
- Removed the previous redirect_to in each method and replaced it with
          render users_api_v0_departments_path
- Ensured the @users variable was set as follows
          @users = @user.org.users.includes(:department)
which is the way @users is set in the users() method in the controller.
The Json was called without this value during rendering.