Update `database.yml.sample`: add separate test db
Prior to this change, `config/database.yml` was configured such that running tests would be executed against the dev db. This would result in the dev db being destroyed after tests. This change ensures that the tests are executed against a separate test db and that the dev db persists while testing.